refactor: CoinView
